refactor(ParamsTab): Restructure ParamsTab components by moving ParamsTabContent, columns, and data to a dedicated ParamsTab directory for improved organization and maintainability
