+++ view/desktop/src/pages/RequestPage/tabs/ParamsTab/cells/index.ts
@@ -1,0 +1,5 @@
+export { ParamInputCell } from "./ParamInputCell";
+export { TemplateInputCell } from "./TemplateInputCell";
+export { TypeSelectCell } from "./TypeSelectCell";
+export { ActionsCell } from "./ActionsCell";
+export { EnabledCheckboxCell } from "./EnabledCheckboxCell";
