+++ view/desktop/src/pages/RequestPage/tabs/ParamsTab/index.ts
@@ -1,0 +1,5 @@
+export { ParamsTabContent } from "./ParamsTabContent";
+export { paramColumns } from "./columns";
+export { queryParamsData, pathParamsData } from "./data";
+export type { ExtendedCellContext } from "./types";
+export * from "./cells";
