+++ view/desktop/src/pages/RequestPage/tabs/ParamsTab/types.ts
@@ -1,0 +1,5 @@
+import { CellContext } from "@tanstack/react-table";
+
+export interface ExtendedCellContext<TData, TValue> extends CellContext<TData, TValue> {
+  focusOnMount?: boolean;
+}
