+++ view/desktop/src/pages/RequestPage/tabs/index.ts
@@ -1,6 +1,6 @@
 export { AuthTabContent } from "./AuthTabContent";
 export { BodyTabContent } from "./BodyTabContent";
 export { HeadersTabContent } from "./HeadersTabContent";
-export { ParamsTabContent } from "./ParamsTabContent";
+export { ParamsTabContent } from "./ParamsTab";
 export { PostRequestTabContent } from "./PostRequestTabContent";
 export { PreRequestTabContent } from "./PreRequestTabContent";
